+++ app/src/main/java/pt/ulisboa/tecnico/cmov/airdesk/entity/ForeignWorkspace.java
@@ -5,7 +5,7 @@ package pt.ulisboa.tecnico.cmov.airdesk.entity;
  */
 public class ForeignWorkspace extends AbstractWorkspace {
 
-    public ForeignWorkspace(String workspaceName, String ownerId, long quota) {
+    public ForeignWorkspace(String workspaceName, String ownerId, double quota) {
         super(workspaceName, ownerId, quota);
     }
 }
